requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

gunicorn>=22.0.0,<23.0.0
uvicorn>=0.23.2,<0.24.0
fastapi>=0.111.1,<0.112.0
psycopg2>=2.9,<3.0
SQLAlchemy>=2.0.31,<2.1.0
sqlmodel>=0.0.20,<0.0.21
alembic>=1.15.2,<1.16.0
pydantic-settings>=2.6.1,<3.0.0
httpx>=0.24.1,<=0.27.0
pytest>=7.4.2,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
pytest-mock>=3.14.0,<3.15.0
email-validator>=2.2.0,<3.0.0
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.1
fastapi-mail>=1.4.1,<1.5.0
Jinja2>=3.1.2,<3.2.0
PyJWT>=2.8.0,<2.9.0
pytz>=2022.1,<2023.0
pandas>=2.1.1,<2.2.0
numpy>=1.26.4,<1.27.0
azure-monitor-opentelemetry>=1.0.0,<2.0.0
python-multipart>=0.0.20
fastapi-mcp>=0.3.3,<=0.3.4
semantic-kernel